Renamed Note component to NoteCard
