--- src/components/organisms/Note/index.ts
@@ -1,1 +1,0 @@
-export * from './Note';
