+++ src/components/organisms/NoteCard/NoteCard.style.tsx
@@ -1,5 +1,5 @@
 import { StyleSheet } from 'react-native';
-import { colors } from '../../../design/tokens/colors';
+import { colors } from '../../../design/tokens';
 
 export const styles = StyleSheet.create({
   card: {
