+++ src/components/organisms/NoteCard/index.ts
@@ -1,0 +1,1 @@
+export * from './NoteCard';
